Refatoração e adição de novos testes.

+++ src/main/java/caixa/ada/service/ContaService.java
@@ -4,6 +4,7 @@ import caixa.ada.DTO.ClienteDTO;
 import caixa.ada.DTO.mapper.ClienteMapper;
 import caixa.ada.exceptions.AgenciaNaoEncontradaException;
 import caixa.ada.exceptions.CepNaoEncontradoOuNuloException;
+import caixa.ada.exceptions.ContaNaoEncontradaException;
 import caixa.ada.model.Agencia;
 import caixa.ada.model.Cliente;
 import caixa.ada.model.Conta;
@@ -32,22 +33,26 @@ public class ContaService {
     }
 
     public void cadastrarConta(ClienteDTO clienteDTO){
-        AgenciaHttp agenciaHttp;
-        try {
-            if (consultaCepHttpService.buscaCep(clienteDTO.getCep()) != null) {
-                agenciaHttp = consultaCepHttpService.buscaCep(clienteDTO.getCep());
-            } else
-                throw new CepNaoEncontradoOuNuloException("CEP inválido ou não encontrado!");
-        } catch (Exception e) {
-            throw new CepNaoEncontradoOuNuloException("CEP inválido ou não encontrado!");
+        AgenciaHttp agenciaHttp = obterAgenciaPorCep(clienteDTO.getCep());
+        if (agenciaHttp.getUf() == null) {
+            throw new AgenciaNaoEncontradaException("Agência não encontrada para o estado informado.");
         }
-        if (agenciaHttp.getUf() != null) {
-            Agencia agencia = agenciaService.buscarUf(agenciaHttp.getUf());
-            Cliente cliente = ClienteMapper.toEntity(clienteDTO);
-            Conta conta = new Conta(agencia, cliente);
-            contaRepository.persist(conta);
-        } else
-            throw new AgenciaNaoEncontradaException("Agencia Não Encontrada! Verifique o CEP informado");
+        Agencia agencia = agenciaService.buscarUf(agenciaHttp.getUf());
+        Cliente cliente = ClienteMapper.toEntity(clienteDTO);
+        Conta conta = new Conta(agencia, cliente);
+        contaRepository.persist(conta);
+    }
+
+    public AgenciaHttp obterAgenciaPorCep(String cep){
+        try {
+            AgenciaHttp agencia = consultaCepHttpService.buscaCep(cep);
+            if (agencia == null) {
+                throw new CepNaoEncontradoOuNuloException("CEP inválido ou não encontrado.");
+            }
+            return agencia;
+        } catch (Exception e) {
+            throw new CepNaoEncontradoOuNuloException("Erro ao consultar o CEP.");
+        }
     }
 
     public List<Conta> getContas() {
@@ -59,10 +64,14 @@ public class ContaService {
     }
 
     public void encerrarConta(Long id) {
-        contaRepository.findById(id).setDataEncerramento();
+        if (this.findById(id) != null) {
+            contaRepository.findById(id).setDataEncerramento();
+        } else throw new ContaNaoEncontradaException("Não encontrada conta para encerrar com esse 'id'.");
     }
 
     public void deletarConta(Long id) {
-        contaRepository.deleteById(id);
+        if (this.findById(id) != null) {
+            contaRepository.deleteById(id);
+        } else throw new ContaNaoEncontradaException("Não encontrada conta para deletar com esse 'id'.");
     }
 }
